put utils of contextual bandit in file contextual_bandit_utils.py
