+++ banditpylib/bandits/__init__.py
@@ -4,6 +4,7 @@ from .ordinary_mnl_bandit import *
 from .linear_bandit import *
 from .thres_bandit import *
 from .contextual_bandit import *
+from .contextual_bandit_utils import *
 
 __all__ = [
     'Bandit', 'OrdinaryBandit', 'LinearBandit', 'Reward', 'MeanReward',
